third_party/blink/web_tests/inspector-protocol/emulation/device-scale-not-persistant-expected.txt

Test that srcset does not use wrong image when override scalefactor and then disabled.
Set deviceScaleFactor: 1
Reloading Page

Page reloaded.

Used Image: /resources/square.png
Set deviceScaleFactor: 2
Reloading Page

Page reloaded.

Used Image: /resources/square200.png
Clear deviceScaleFactor
Reloading Page

Page reloaded.

Current image src equal initial image: Yes
